collect_device/requirements.txt
fastapi
uvicorn
psutil
py-cpuinfo
pyyaml
gitpython